Mark non-independent DESTINY+ predictions unfalsifiable (v1.0.2)
Most severe finding of the ecosystem-wide Gamma-circularity review:
GAMMA_PHAETHON=0.165 has NO cited physical derivation -- its original
comment described it only as "target: between Amazon (0.116) and AMOC
(0.251)", i.e. chosen to occupy a plausible slot in the shared
cross-package Gamma ordering, not derived from any measured or
modelled property of asteroid 3200 Phaethon. This value was presented
as DESTINY+ Prediction #1 with an uncertainty band, framed as
falsifiable against the 2029 flyby. Prediction #4 (H*) is directly
computed from #1 and #3, so it is not independent either. Both now
marked falsifiable=False in destiny_predictions.py (the dataclass
already had this field, never previously set). Prediction #3's
"Beta-to-Gamma bridge (P32)" citation traces to beta-clustering-utac's
own unverified claim -- flagged separately in FOLLOWUP_TICKETS.md.
Predictions #5-47 were not individually re-audited.

Updated constants.py, destiny_predictions.py, cli.py (destiny-report
now shows a Falsifiable column), README.md, .zenodo.json. Also fixed 3
pre-existing mypy --strict errors found while touching these files.

+++ src/phaethon_chimera/frustrated_utac.py
@@ -88,7 +88,7 @@ class FrustratedUTAC:
             t_end = n_orbits * self.period_days
 
         steps = int((t_end - t_start) / dt)
-        times = np.linspace(t_start, t_end, steps)
+        times = np.linspace(t_start, t_end, steps, dtype=np.float64)
         H = np.zeros(steps)
         H[0] = max(1e-6, H0)
